Add graph construction logic to PathFinder
A graph construction logic has been added to the PathFinder.js file, mapping the connections between the different cells on the board. Additionally, a corresponding test has been added to pathFinder.test.js to ensure accurate creation and initialization of this graph. This implementation will allow the program to map the path of the knight more efficiently.