Refactor unit tests and simplify path finding function
A huge portion of unit tests in 'pathFinder.test.js' was removed, focusing on testing the path lengths, which were slightly modified as well. In 'pathFinder.js', the function 'getMinimumMoves' was simplified removing a chunk of code that populated the graph if a target does not exist. This results in a less complex and more maintainable code.